requirements.txt
flask==3.0.2
pymongo==4.6.1
flask-cors==4.0.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
flask-limiter==3.5.0
bcrypt==4.1.2
marshmallow==3.20.2
flask-marshmallow==0.15.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0 